% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asset_id}
\alias{asset_id}
\title{Create asset id}
\usage{
asset_id(tipo, emisora, serie)
}
\arguments{
\item{tipo}{character vector of tipo}

\item{emisora}{character vector of emisora}

\item{serie}{character vector of serie}
}
\value{
the unique id of the asset
}
\description{
Creates a unique if from the tipo, emisora, and serie of an asset. Input may be vector but must all have the same length;
if they are not, it will not neccesarily yield an error but undesired results may occur.
}
\examples{
asset_id("M", "BONOS", "191211")
}

