% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posiciones.R
\name{clean_position}
\alias{clean_position}
\title{Remove uncommon holdings from position data}
\usage{
clean_position(position)
}
\arguments{
\item{position}{A data.frame object containing holdings information. It is assumed to be generated using \code{\link{get_position}}}
}
\value{
A data.frame with the holdings after removing id's that have zero median within a strategy
}
\description{
Computes the median percentage from total within each strategy and removes holdings from position data that have median zero.
}
\details{
Computes the median percentage from total within each strategy and removes holdings from position data that have median zero.
It is the opposite of \code{\link{find_rare_position}}. Sometimes even after removing blocked holdings using \code{\link{remove_blocked}} there
are holdings that belong to client decision and not our investment strategies (the position has not yet been blocked). This method
can be used to further clean the data by removing elements that are uncommon within a strategy
}
\examples{
clean_position(get_position("2015-12-31"))
}
\seealso{
\code{\link{find_rare_position}}
}
