% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_seq}
\alias{date_seq}
\title{Convenient date sequences}
\usage{
date_seq(period, by = 1)
}
\arguments{
\item{period}{A character of the form 'date1/date2', where the date1 and date2 are converted do Date objects using \code{lubridate}.}

\item{by}{an integer specifying the space between dates in the sequence}
}
\value{
A sequence of dates
}
\description{
A wrapper for seq.Date in more convenient form
}
\details{
A convenient wrapper for seq.date that does not requite tu input Dates and accepts other date formats using \code{ymd} from \code{lubridate}.
The syntax is inspired on the indexing format of xts objects.
}
\examples{
date_seq("20151231/")
date_seq("20151231/20161130", by = 2)
}

