% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{dly_ret}
\alias{dly_ret}
\title{Daily returns for wide-form price data}
\usage{
dly_ret(df, halflife, roll = 1, accumulate = FALSE, date_col = NULL)
}
\arguments{
\item{df}{A data frame or xts with price data. If a data frame is given then the first column of date type will as
the index unless it is manually set in date_col}

\item{halflife}{To be implemented}

\item{roll}{Rolling returns, "1 = no rolling"}

\item{date_col}{To be implemented}

\item{type}{To be implemented}
}
\value{
An object of the same class of df with the returns.
}
\description{
Daily returns for wide-form price data
}
\examples{
tq_get_wide(c("SPY", "SHV")) \%>\% dly_ret()
}
