% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_prices}
\alias{fill_prices}
\title{Fill data frame or xts}
\usage{
fill_prices(df, fill_missing_init = TRUE, date_col = NULL)
}
\arguments{
\item{df}{data.frame or xts with data to fill}

\item{fill_missing_init}{A boolean indicating whether starting missing values should be copied to first actual value}
}
\value{
Filled data frame
}
\description{
Fills missing data in convenient form for Finance
}
\examples{
fill_prices(tq_get_wide("SPY,MXN=X"))
}
