% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composiciones.R
\name{get_carteramodelo}
\alias{get_carteramodelo}
\title{Fetch cartera modelo}
\usage{
get_carteramodelo(fecha, contrato = NULL)
}
\arguments{
\item{fecha}{A vector of dates that can be of class character or Date but is expected to be in the format aaaa-mm-dd}
}
\value{
A data.frame with the cartera modelo (composition) of the given contratos in the selected dates.
}
\description{
Construct a mysql query for the selected dates it to the sql database 'carteramodelo'
}
\details{
This function will return a data.frame with cartera modelo (strategy objetive composition) information from one or several dates
}
\examples{
fecha <- seq(as.Date("2015-12-31"), as.Date("2016-01-31"), by = 1)
get_carteramodelo(fecha)
}

