% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posiciones.R
\name{get_contrato_returns}
\alias{get_contrato_returns}
\title{returns of the total valuation per contrato}
\usage{
get_contrato_returns(position, cumulative = FALSE)
}
\arguments{
\item{position}{A position data.frame object assumed to be generated using \code{\link{get_position}}}

\item{cumulative}{A boolean indicated whether returns are cumulative or daily. Defaults to FALSE}
}
\value{
A \code{\link{xts}} (time series) object indexed by the date and with each column representing the daily return of the contrato.
Returns are give in the form of valuation ratio (final_price/initial_price)
}
\description{
Create return data per contract for each contrato
}
\details{
This function takes a position data.frame and uses it to compute the return of the contrato by taking the percent change
of each date's valuation and its previous valuation indicated by the column named \code{tot}.
}
\examples{
pos <- get_position(c("2015-12-31", "2016-01-04", "2016-01-05"), contrato = c("25774", "25773"))
get_position_returns(pos)
}

