% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercados.R
\name{get_mercados}
\alias{get_mercados}
\title{Fetch asset information data from Amazon}
\usage{
get_mercados(id = NULL, where = NULL)
}
\value{
A data.frame with the prices of the given ids in the selected dates.
}
\description{
Construct a mysql query for the selected dates and ids from the Amazon mercados database, table valores
}
\examples{
get_mercados()
#' get_mercados(c("1ispy_", "1ishv_"))
}
