% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posiciones.R
\name{get_position}
\alias{get_position}
\title{Fetch holdings data}
\usage{
get_position(date, contrato = NULL)
}
\arguments{
\item{date}{A vector of dates that can be of class character or Date but is expected to be in the format aaaa-mm-dd}

\item{contrato}{A character or integer vector with the number of contratos}
}
\value{
A data.frame with the holding of the given contratos in the selected dates. If no contrato is given, it will return the information of every contrato.
}
\description{
Construct a mysql query for the selected dates and contrato and sends it to the sql database posiciones
}
\details{
This function will return a data.frame with holdings information from one or several dates.
When no contrato is given, it will fetch the information of every contrato.
}
\examples{
get_position("2015-12-31")
get_position("2015-12-31", contrato = "25774")
}

