% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precios.R
\name{get_prices}
\alias{get_prices}
\title{Fetch price data from Amazon}
\usage{
get_prices(fecha, id = NULL)
}
\arguments{
\item{fecha}{A vector of dates that can be of class character or Date but in the format aaaa-mm-dd}

\item{id}{A character vector with id list to query}
}
\value{
A data.frame with the prices of the given ids in the selected dates.
}
\description{
Construct a mysql query for the selected dates and ids from the Amazon prices database
}
\details{
This function will return a data.frame with a column per id and date rows
}
\examples{
get_prices(date_seq("2015-12-31/"), c("1ispy_", "1ishv_"))
}
