% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keymapping}
\alias{keymapping}
\title{Map a variable}
\usage{
keymapping(x, origin, target, replace_missing = FALSE)
}
\arguments{
\item{x}{The vector that will be transformed; x must have values in \code{origin}.}

\item{origin}{A vector}

\item{target}{A vector of the same length as origin}

\item{replace_missing}{Indicates whether the values of \code{x} that are not found \code{origin} should be changed.
If \code{replace_missing} is given a value other than \code{FALSE} then this value will be used as replacement for the
values of \code{x} not found in \code{origin}.}
}
\value{
A transformed vector
}
\description{
Sometimes we need to map one vector into another. For inhouse data it can be useful when defining categories for
carteras modelo, id's or emisoras. This function maps \code{origin -> target} for every value of x.
}
\examples{
keymapping(1:6, c(2,3), c(-1, -2))
keymapping(1:6, c(2,3), c(-1, -2), replace_missing = NA)
}
