% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{OptimistFunctions}
\alias{OptimistFunctions}
\alias{optimist_mean}
\alias{optimist_var}
\alias{optimist_cov}
\title{Statistical functions with halflife for financial analysis}
\usage{
optimist_mean(x, halflife = length(x), type = c("arithmetic", "geometric"))

optimist_var(x, halflife = length(x))

optimist_cov(x, halflife = nrow(as.matrix(x)), cor = FALSE)
}
\arguments{
\item{halflife}{the value for time-weighting. See \link{optimist_wt}.}

\item{object}{x}
}
\value{
mean, variance and convariance matrix of \code{x} with halflife \code{h}.
}
