% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precios.R
\name{price_ids}
\alias{price_ids}
\title{Id list}
\usage{
price_ids(fecha = NULL)
}
\arguments{
\item{fecha}{In case they must only be selected from a date (faster)}
}
\value{
A character vector the full list of unique ids
}
\description{
Generate the full list of available ids from Amazon
}
\examples{
price_ids()
price_ids("2015-12-31")
}

