% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composiciones.R, R/mercados.R, R/posiciones.R,
%   R/precios.R, R/returns.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{tq_get}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%>\%}}}

  \item{tidyquant}{\code{\link[tidyquant]{tq_get}}}
}}

