% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posiciones.R
\name{remove_blocked}
\alias{remove_blocked}
\title{Remove blocked holdings}
\usage{
remove_blocked(position, blocked)
}
\arguments{
\item{position}{A data.frame containing the full holdings information.
It is assumed to be generated with the function \code{\link{get_position}}}

\item{blocked}{A data.frame containing the full holdings information
It is assumed to be generated with the function \code{\link{get_blocked_position}}}
}
\value{
A data.frame with the holdings that are not blocked (those that belong to the strategies and not
the client decisions and thus can be rebalanced)
}
\description{
Remove all the holdings that are blocked from holdings data
}
\details{
This function will match the entries of a data frame representing blocked holdings and remove the corresponding titles
and valuation from another data frame containing the full position
}
\examples{
date <- "2016-08-31"
pos <- get_position(date)
bl <- get_blocked_position(date)
unbl <- rm_blocked(pos, bl)
head(unbl)
}

