% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rollapply.xts}
\alias{rollapply.xts}
\title{rollapply xts as it should be}
\usage{
\method{rollapply}{xts}(data, ...)
}
\arguments{
\item{data}{the data to be used (representing a series of observations).}

\item{...}{arguments to pass to \code{\link{rollapply.zoo}}}
}
\value{
Corces xts to zoo, then applies \code{\link{rollapply.zoo}} of \code{\link{zoo}} package and transforms back to xts
}
\description{
Overwrites rollapply.xts and uses rollapply.zoo for xts data
}
\details{
The function transforms
}
