% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rollProd}
\alias{rollProd}
\title{Rolling product for xts}
\usage{
rollProd(x, roll = 1)
}
\arguments{
\item{x}{what to roll}

\item{roll}{rolling window}
}
\value{
rolled product of x
}
\description{
Wrapper of RcppRoll::roll_prod
}
