% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spl}
\alias{spl}
\title{Split string  by symbol}
\usage{
spl(x, sep = ",")
}
\arguments{
\item{x}{single string}

\item{sep}{symbol to split, defaults to ','}
}
\value{
x splitted and unlisted into vector
}
\description{
Splits a string by a symbol, trims whitespace and unlists
}
\examples{
spl("SPY,SHV")
}
\seealso{
\link{str_split}
}
