% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posiciones.R
\name{summarize_position}
\alias{summarize_position}
\title{Summary statistics for holdings data}
\usage{
summarize_position(position, groups = c("fecha", "carteramodelo", "id"), ...)
}
\arguments{
\item{position}{A data.frame object containing holdings information. It is assumed to be generated using \code{\link{get_position}}}

\item{groups}{character vector of the names of the variables that will be used with grouping and computing statistics
Variable names must be quoted (an inconvenience used to avoid confusion with the commands passed to \code{\link{summarize}}).
If no grouping is desired then it can be set to NULL}

\item{...}{Commands to be passed to \code{\link{summarize}} from \code{\link{dplyr}} for computing statistics within groups
#' @details The function will group by investment strategy (carteramodelo), date and asset id and compute
basic summary statistics and any other command specified by the user.}
}
\value{
A data.frame object with the summary statistics. By default it returns a data.frame indicating the group and columns
\itemize{
  \item{\strong{sum_mon}:} {The sum of the valuation of the group}
  \item{\strong{perc_median}:} {The median of the percentage of the total valuation}
  \item{\strong{perc_mean}:} {The mean of the percentage of the total valuation}
  \item{\strong{perc_sd}:} {The standard deviation of the percentage of the total valuation}
  \item{}{The result of additional commands supplied by the user}
}
}
\description{
Compute basic summary statistics for holdings data for each strategy, date and id
}
\examples{
position <- get_position(fecha = c("2016-07-28", "2016-07-29"))
summarize_position(position, groups = c("fecha", "carteramodelo", "id"), count = length(contrato))
summarize_position(position, groups = NULL)
position <- position \%>\% map_id()
}

