% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{tq_get_wide}
\alias{tq_get_wide}
\title{Yahoo Data Wide}
\usage{
tq_get_wide(x, what = "close", ...)
}
\arguments{
\item{x}{A single character string, a character vector or data frame
representing a single or multiple stock index, stock symbol, metal symbol, currency combination}

\item{...}{Additional parameters passed to the appropriate quantmod function. Common optional parameters include:
\itemize{
    \item \code{from} A character string representing a start date in YYYY-MM-DD format
    \item \code{to} A character string representing a end date in YYYY-MM-DD format.
}}

\item{type}{a single \code{character} element from the following options
\code{"open", "close", "high", "low", "volume", "adjusted"}. Defaults to "close".}
}
\value{
A data frame with the first column as the date and each other column representing a stock and each cell its price.
}
\description{
Get yahoo price data in wide form for a vector of stocks.
}
\examples{
tq_get_wide(c("SPY", "SHV"))
tq_get_wide(c("SPY", "SHV"), what = "volume", from = "2015-12-31")
}
\seealso{
\link{tq_get}, \link{tq_spread}
}
