% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{tq_spread}
\alias{tq_spread}
\title{Spread tq_get prices}
\usage{
tq_spread(tbl, as_xts = FALSE, what = "close", ...)
}
\arguments{
\item{tbl}{The results from using \code{tq_get} with the option \code{get = "stock.prices"}}

\item{what}{a single \code{character} element from the following options
\code{"open", "close", "high", "low", "volume", "adjusted"}. Defaults to "close".}

\item{...}{Additional parameters passed to the appropriate quantmod function. Common optional parameters include:
\itemize{
    \item \code{from} A character string representing a start date in YYYY-MM-DD format
    \item \code{to} A character string representing a end date in YYYY-MM-DD format.
}}
}
\value{
A data frame with the first column as the date and each other column representing a stock and each cell its price.
}
\description{
A method to use with \link{tq_get} from the \link{tidyquant} package.
}
\details{
To use only after \code{tq_get}.
}
\seealso{
\link{tq_get}, \link{tq_get_wide}
}
